refactor: use InitializationConfiguration from CSL
