+++ qa/acceptance-tests/src/test/java/io/camunda/it/logout/BasicAuthLogoutIT.java
@@ -7,8 +7,8 @@
  */
 package io.camunda.it.logout;
 
-import static io.camunda.security.configuration.InitializationConfiguration.DEFAULT_USER_PASSWORD;
-import static io.camunda.security.configuration.InitializationConfiguration.DEFAULT_USER_USERNAME;
+import static io.camunda.security.api.model.config.initialization.InitializationConfiguration.DEFAULT_USER_PASSWORD;
+import static io.camunda.security.api.model.config.initialization.InitializationConfiguration.DEFAULT_USER_USERNAME;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.qa.util.cluster.TestCamundaApplication;
