+++ qa/util/src/main/java/io/camunda/qa/util/auth/Authenticated.java
@@ -8,7 +8,7 @@
 package io.camunda.qa.util.auth;
 
 import io.camunda.client.CamundaClient;
-import io.camunda.security.configuration.InitializationConfiguration;
+import io.camunda.security.api.model.config.initialization.InitializationConfiguration;
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
